requirements.txt
# Job Application Tracker - Dependencies
# Core framework
pocketflow>=0.1.0

# Google API integration
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# HTTP client and async support
httpx>=0.25.0
aiohttp>=3.8.0
requests>=2.31.0
asyncio-throttle>=1.0.0

# Data validation and serialization
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0

# LLM integration
openai>=1.0.0

# Data processing
pyyaml>=6.0.0

# CLI and terminal UI
rich>=13.0.0
click>=8.0.0

# Email processing
email-validator>=2.0.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16

# Date/time handling
python-dateutil>=2.8.0

# File I/O
aiofiles>=23.0.0

# Testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
responses>=0.23.0
freezegun>=1.2.0

# Code quality
ruff>=0.1.0
mypy>=1.5.0

# Type stubs
types-PyYAML>=6.0.0
types-python-dateutil>=2.8.0

# Development tools
black>=23.0.0
pre-commit>=3.0.0